+++ cloud-sql/sql-server/sqlalchemy/requirements.txt
@@ -1,7 +1,7 @@
 Flask==2.1.0
-SQLAlchemy==1.4.36
+google-cloud-sql-python-connector==0.6.1
 python-tds==1.11.0
-sqlalchemy-pytds==0.3.4
 pyodbc==4.0.32
 pyopenssl==22.0.0
-
+SQLAlchemy==1.4.36
+sqlalchemy-pytds==0.3.4
